requirements.txt
streamlit
tensorflow
scikit-learn
pillow
numpy
pandas
nltk
matplotlib
seaborn